requirements.txt
Pillow
google-auth-oauthlib
google-api-python-client
google-auth
google-auth-oauthlib
